refactor(fastapi-rest-tpl): reorganize folder structure to start using clean architecture
