--- backend/ms-fastapi-rest-tpl/CHANGELOG.md
@@ -1,19 +1,0 @@
-## 1.0.0 (2025-07-09)
-
-* fix(cicd): modified github release step to adjust asset building (#82) ([fc0b6fb](https://github.com/infojasyrc/app-developer-experience/commit/fc0b6fb)), closes [#82](https://github.com/infojasyrc/app-developer-experience/issues/82)
-* fix(cicd): modified github release step to adjust asset building and create tag (#83) ([bf38ef6](https://github.com/infojasyrc/app-developer-experience/commit/bf38ef6)), closes [#83](https://github.com/infojasyrc/app-developer-experience/issues/83)
-* docs(fastapi-rest-tpl): added more commands in readme for fastapi rest template (#72) ([9df73d2](https://github.com/infojasyrc/app-developer-experience/commit/9df73d2)), closes [#72](https://github.com/infojasyrc/app-developer-experience/issues/72)
-* docs(fastapi-rest-tpl): added project description in pyproject.toml (#74) ([2412b58](https://github.com/infojasyrc/app-developer-experience/commit/2412b58)), closes [#74](https://github.com/infojasyrc/app-developer-experience/issues/74)
-* docs(fastapi-rest-tpl): added project structure in readme (#78) ([936d811](https://github.com/infojasyrc/app-developer-experience/commit/936d811)), closes [#78](https://github.com/infojasyrc/app-developer-experience/issues/78)
-* docs(fastapi-rest-tpl): update changelog ([2e739d7](https://github.com/infojasyrc/app-developer-experience/commit/2e739d7))
-* docs(fastapi-rest-tpl): update changelog ([5f4f101](https://github.com/infojasyrc/app-developer-experience/commit/5f4f101))
-* docs(fastapi-rest-tpl): update changelog ([1dd5761](https://github.com/infojasyrc/app-developer-experience/commit/1dd5761))
-* docs(ms-fastapi-rest-tpl): added main README (#70) ([b161f03](https://github.com/infojasyrc/app-developer-experience/commit/b161f03)), closes [#70](https://github.com/infojasyrc/app-developer-experience/issues/70)
-* feat(cicd): added library, github workflow and installed act to run local github action (#68) ([1d076e1](https://github.com/infojasyrc/app-developer-experience/commit/1d076e1)), closes [#68](https://github.com/infojasyrc/app-developer-experience/issues/68)
-* feat(cicd): modified release asssets to use only changed project (#81) ([c2b8b72](https://github.com/infojasyrc/app-developer-experience/commit/c2b8b72)), closes [#81](https://github.com/infojasyrc/app-developer-experience/issues/81)
-* feat(fastapi-rest-tpl): use dev sufix for development and update compose configuration (#76) ([80a1e77](https://github.com/infojasyrc/app-developer-experience/commit/80a1e77)), closes [#76](https://github.com/infojasyrc/app-developer-experience/issues/76)
-* ms-fastapi-rest-tpl/feat: upgrade dependencies with poetry (#69) ([34953f7](https://github.com/infojasyrc/app-developer-experience/commit/34953f7)), closes [#69](https://github.com/infojasyrc/app-developer-experience/issues/69)
-* chore: move python ms using fast api (#17) ([35542eb](https://github.com/infojasyrc/app-developer-experience/commit/35542eb)), closes [#17](https://github.com/infojasyrc/app-developer-experience/issues/17)
-
-
-
