+++ backend/ms-fastapi-rest-tpl/src/core/constants.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class Environment(str, Enum):
+    """Enum representing the different environments."""
+    DEVELOPMENT = "development"
+    TESTING = "testing"
+    PRODUCTION = "production"
