--- backend/ms-fastapi-rest-tpl/src/core/environment.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-class Environment(str, Enum):
-    development = "development"
-    testing = "testing"
-    production = "production"
